Fix Bug: Qualities exceeding one.
Qualities were exceeding one due to floating-point errors. The distance
between the query point and the optimal solution as well as between the
candidate and the query point were calculated. If the candidate solution
was the optimal solution, then it could still occur that floating-point
errors produced slightly different distances.

For small distances this error was exascerbated when calculating the
quality, leading to, in theory, qualities exceeding 1 arbitrarily. The
solution was to pre-compute all distances and use them when calculating
the quality. This way, qualities are guaranteed to be equal to 1 for
optimal solutions.